LayoutTests/imported/w3c/web-platform-tests/focus/iframe-contentwindow-focus-with-same-as-top-intermediate-frame-expected.txt


FAIL Check result assert_equals: Check log expected "outeronload;outeractivelement:body;starttest;middleactivelement:body;callfocus;middleactivelement:iframe;innerfocus;middletimer;middleactivelement:iframe;outeractivelement:iframe;" but got "outeronload;outeractivelement:body;starttest;middleactivelement:body;callfocus;middleactivelement:body;middletimer;middleactivelement:body;outeractivelement:body;"
